+++ projects/core/src/cms/store/effects/component.effect.ts
@@ -9,12 +9,10 @@ import { CmsComponent } from '../../../model/cms.model';
 import { PageContext } from '../../../routing/index';
 import { SiteContextActions } from '../../../site-context/store/actions/index';
 import { bufferDebounceTime } from '../../../util/buffer-debounce-time';
-import {
-  makeErrorSerializable,
-  serializePageContext,
-} from '../../../util/serialization-utils';
+import { makeErrorSerializable } from '../../../util/serialization-utils';
 import { withdrawOn } from '../../../util/withdraw-on';
 import { CmsComponentConnector } from '../../connectors/component/cms-component.connector';
+import { serializePageContext } from '../../utils/cms-utils';
 import { CmsActions } from '../actions/index';
 
 @Injectable()
